+++ CHANGELOG.md
@@ -11,7 +11,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 [bc]: https://github.com/dogmatiq/.github/blob/main/VERSIONING.md#changelogs
 
-## [Unreleased]
+## [0.15.1] - 2026-02-07
 
 ### Added
 
@@ -371,6 +371,7 @@ This eliminates the need for duplicating all of the generic algorithms (such as
 [0.13.0]: https://github.com/dogmatiq/persistencekit/releases/tag/v0.13.0
 [0.14.0]: https://github.com/dogmatiq/persistencekit/releases/tag/v0.14.0
 [0.15.0]: https://github.com/dogmatiq/persistencekit/releases/tag/v0.15.0
+[0.15.1]: https://github.com/dogmatiq/persistencekit/releases/tag/v0.15.1
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD